chore(core): rename releaseActivity to releaseEvents
